feat: add navibar Amenity section
