+++ frontend/src/app/app.tsx
@@ -14,13 +14,14 @@ import TopReviews from 'src/pages/Detail/TopReviews';
 import Reviews from 'src/pages/Reviews/Reviews';
 import ProtectedRoute from 'src/pages/ProtectedRouter/ProtectedRouter';
 import Search from 'src/pages/Search/Search';
+import Amenity from 'components/Amenity/Amenity';
 
 const App = () => {
   const { id, nickname, email } = useAppSelector(selectAuth);
 
   const [detailNavigation] = useState([
     { id: 'rooms', href: '', content: '객실' },
-    { id: 'test', href: 'test', content: '편의시설' },
+    { id: 'amenities', href: 'amenities', content: '편의시설' },
     { id: 'topReviews', href: 'topReviews', content: '후기' },
   ]);
 
@@ -42,6 +43,7 @@ const App = () => {
           <Route path="/reviews" element={<Reviews />} />
           <Route path="/detail" element={<Detail list={detailNavigation} />}>
             <Route index element={<Rooms />} />
+            <Route path="amenities" element={<Amenity />} />
             <Route path="topReviews" element={<TopReviews />} />
           </Route>
         </Routes>
