+++ frontend/src/components/Amenity/Amenity.tsx
@@ -1,6 +1,8 @@
+import { useState, useEffect } from 'react';
 import Review from 'components/Review/Review';
 import { faStar, faThumbsUp } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { getHotelInfo } from 'src/utils/requests';
 
 import {
   AmenityWrapper,
@@ -13,8 +15,35 @@ import {
 } from './Amenity.style';
 
 import { FeaturesWrapper, FeaturesTitle, FeaturesBox, FeaturesBoxTitle, FeatureList } from './AmenityFeatures.style';
+const settingHotelOverview = (body: object[]) => {
+  interface Overview {
+    title?: string;
+    type?: string;
+    content: string[];
+  }
+  let sectionSetting: Overview = {
+    title: body.amenities[0].heading,
+    type: body.amenities[0].listItems[0].heading,
+    content: body.amenities[0].listItems[0].listItems,
+  };
 
-const Amenity = ({ overviews }) => {
+  const hotelOverview: object[] = [
+    body.overview.overviewSections[0],
+    sectionSetting,
+    body.overview.overviewSections[1],
+  ];
+
+  return hotelOverview;
+};
+const Amenity = () => {
+  const [overviews, setOverviews] = useState<object[]>([]);
+  useEffect(() => {
+    const requestbody = async () => {
+      const resoponseInfo = await getHotelInfo(hotelId);
+      setOverviews(settingHotelOverview(resoponseInfo));
+    };
+    requestbody();
+  }, []);
   const mockReviewList = [
     { id: '1', content: '편리함', count: 10 },
     { id: '2', content: '쇼핑', count: 11 },
@@ -22,49 +51,6 @@ const Amenity = ({ overviews }) => {
     { id: '4', content: '교통편 옵션', count: 27 },
     { id: '5', content: '다이닝 옵션', count: 32 },
     { id: '6', content: '걷기좋음', count: 22 },
-  ];
-
-  const overReiwMock = [
-    {
-      id: '0',
-      title: ' 주요 편의 시설',
-      type: 'HOTEL_FEATURE',
-      content: [
-        '462개의 금연 객실',
-        '매일 하우스키핑',
-        '5 개 레스토랑 및 바/라운지',
-        '실내 수영장',
-        '피트니스 센터',
-        '주차 대행',
-        '스파 서비스',
-        '비즈니스 센터',
-        '공항 셔틀',
-        '24시간 운영 프런트 데스크',
-        '프런트 데스크 금고',
-        '스낵바/델리',
-        '무료 WiFi 및 무료 주차',
-      ],
-    },
-    {
-      id: '1',
-      title: '기타 편의 시설',
-      type: 'HOTEL_FEATURE_ETC',
-      content: ['객실 내 온도 조절기', '미니바', '에스프레소 메이커', '목욕가운', '슬리퍼', '다리미/다리미판'],
-    },
-    {
-      id: '2',
-      title: '주변 명소',
-      type: 'LOCATION_SECTION',
-      content: [
-        '중구에 위치',
-        '서울특별시청(걸어서 4분 거리)',
-        '롯데백화점(걸어서 5분 거리)',
-        '서울광장(걸어서 5분 거리)',
-        '한국은행 화폐박물관(걸어서 3분 거리)',
-        '눈 스퀘어(걸어서 3분 거리)',
-        '한국 우표박물관(걸어서 4분거리)',
-      ],
-    },
   ];
 
   return (
